manage bootstrap using npm
This effectively reverts the approach of af8c522
keeping the spirit I see behind d2823ab
that is, managing twbs using npm will be
easier to maintain.

Redistributing the minified file prevents
another repo-moved fiasco like #16, while listing
bootstrap as a devDependency means that developers
will be able to peruse the same version of the
unminifed source without it being checked in to
this repository.