Allow passing arguments to the open event listener
Arguments can now be passed from the ContextMenu.open() function to the
function bound to listen to the “open” event. This allows us to, for
example, set the options available in the context menu based on the
mouse event emitted on the context menu click or the element clicked
on.